/!\ unstable : replace old search loop by a search loop with integrated restarts add some solver configuration class (limit, restart)
git-svn-id: svn://svn.code.sf.net/p/choco/code/trunk@1182 5779b9a9-2447-0410-887e-bb98d57c6de4

+++ choco-kernel/src/main/java/choco/kernel/solver/search/restart/LubyRestartStrategy.java
@@ -10,7 +10,7 @@ public final class LubyRestartStrategy  extends AbstractRestartStrategy {
 
 
 	public LubyRestartStrategy(int scaleFactor,
-			double geometricalFactor) {
+			int geometricalFactor) {
 		super("LUBY", scaleFactor, geometricalFactor);
 	}
